asdasdasawe

+++ phishing/views.py
@@ -3,7 +3,10 @@ from django.urls import reverse_lazy
 from django.contrib.auth.decorators import login_required
 from .forms import UrlsForm, MensajeForm, ProxyForm, Search
 from .models import Url, Correo
-from .phishing import verifica_urls, archivo_texto, monitorea_url
+from .phishing import (
+    verifica_urls, archivo_texto, monitorea_url,
+    whois, archivo_comentarios, archivo_hashes
+)
 from .correo import genera_mensaje, manda_correo, obten_asunto, obten_mensaje
 from django.views.generic import TemplateView
 from django.template import loader
@@ -13,6 +16,10 @@ from django.core.exceptions import MultipleObjectsReturned
 from django.conf import settings
 from shutil import copyfile
 import os
+from .reporte import (
+    cuenta_urls, urls_activas, urls_inactivas, urls_redirecciones,
+    urls_entidades, urls_titulos, urls_dominios, urls_paises
+)
 
 @login_required(login_url=reverse_lazy('login'))
 def monitoreo(request):
@@ -22,11 +29,15 @@ def monitoreo(request):
     return render(request, 'monitoreo.html')
 
 def rmimg(img):
+    if img is None:
+        return
     f = os.path.join(settings.MEDIA_ROOT, img)
     if os.path.exists(f):
         os.remove(f)
 
-def cpimg(img, img2):    
+def cpimg(img, img2):
+    if img is None or img2 is None:
+        return
     f = os.path.join(settings.BASE_DIR, img[1:])
     f2 = os.path.join(settings.BASE_DIR, img2[1:])
     if os.path.exists(f):
@@ -41,10 +52,13 @@ def monitoreo_id(request, pk):
         'url': url,
         'dominio': url.dominio.captura_url
     }
-    captura_old = '%s_monitoreo.png' % url.captura_url[:url.captura_url.rindex('.')]
+    if url.captura_url is None:
+        captura_old = None
+    else:
+        captura_old = '%s_monitoreo.png' % url.captura_url[:url.captura_url.rindex('.')]
     if not url is None:
         datos = {
-            'de': settings.CORREO_USR,
+            'de': settings.CORREO_DE,
             'para': ', '.join([x.correo for x in url.correos.all()]),
             'asunto': obten_asunto(url),
             'mensaje': obten_mensaje(url)
@@ -54,12 +68,26 @@ def monitoreo_id(request, pk):
             if request.POST.get('boton-curl'):
                 proxy_form = ProxyForm(request.POST)
                 if proxy_form.is_valid():
-                    servidor = proxy_form.cleaned_data['servidor']
-                    puerto = proxy_form.cleaned_data['puerto']
+                    http = proxy_form.cleaned_data['http']
+                    https = proxy_form.cleaned_data['https']
                     tor = proxy_form.cleaned_data['tor']
+                    proxies = proxy_form.cleaned_data['proxy']
                     proxy = None
                     if tor:
                         proxy = {'http':  'socks5://127.0.0.1:9050', 'https': 'socks5://127.0.0.1:9050'}
+                    elif http or https:
+                        proxy = {}
+                        if http:
+                            proxy['http'] = http
+                        if https:
+                            proxy['https'] = https
+                    elif not proxies is None and (not proxies.http is None or
+                                                  not proxies.http is None):
+                        proxy = {}
+                        if not proxies.http is None:
+                            proxy['http'] = proxies.http
+                        if not proxies.https is None:
+                            proxy['https'] = proxies.https                    
                     cpimg(url.captura_url, captura_old)
                     old = {
                         'captura_url': captura_old,
@@ -84,6 +112,22 @@ def monitoreo_id(request, pk):
                         x.reportado = True
                         x.save()
                     rmimg(captura_old)
+                    context = {
+                        'url': url,
+                        'de': de,
+                        'para': para,
+                        'asunto': asunto,
+                        'mensaje': mensaje,
+                        'captura': url.captura_url
+                    }
+                    return render(request, 'monitoreo_exito.html', context)
+            elif request.POST.get('boton-saltar'):
+                for x in Url.objects.filter(url=url.url):
+                    x.reportado = True
+                    x.save()
+                return redirect('monitoreo')
+            elif request.POST.get('boton-siguiente'):
+                return redirect('monitoreo')
     context['mensaje_form'] = mensaje_form
     context['proxy_form'] = proxy_form
     return render(request, 'monitoreo_id.html', context)
@@ -94,13 +138,45 @@ def valida_urls(request):
         form = UrlsForm(request.POST)
         if form.is_valid():
             urls = form.cleaned_data['urls']
-            verifica_urls([x.strip() for x in urls.split('\n') if x.strip()], None, False)
-            # return redirect('post_detail', pk=post.pk)
+            sitios = verifica_urls([x.strip() for x in urls.split('\n') if x.strip()], None, False)
+            activas = urls_activas(sitios)
+            inactivas = urls_inactivas(sitios)
+            redirecciones = urls_redirecciones(sitios)
+            context = {
+                'urls_total': cuenta_urls(sitios),
+                'num_urls_activas': len(set([x.url for x in activas])),
+                'num_urls_inactivas': len(set([x.url for x in inactivas])),
+                'num_urls_redirecciones': len(set([x.url for x in redirecciones])),
+                'entidades': urls_entidades(sitios),
+                'titulos': urls_titulos(sitios),
+                'dominios': urls_dominios(sitios),
+                'paises': urls_paises(sitios),
+                'activas': activas,
+                'inactivas': inactivas,
+                'redirecciones': redirecciones
+            }
+            return render(request, 'reporte_urls.html', context)
     else:
         form = UrlsForm()
     return render(request, 'valida_urls.html', {'form': form})
 
 message2 = ""
+
+@login_required(login_url=reverse_lazy('login'))
+def url_detalle(request, pk):
+    url = get_object_or_404(Url, pk=pk)
+    comentarios = archivo_comentarios(url)
+    hashes = archivo_hashes(url)
+    wi = whois(url.ip)
+    wi_dom = whois(url.dominio.dominio)
+    context = {
+        'url': url,
+        'comentarios': comentarios,
+        'hashes': hashes,
+        'whois': wi,
+        'whois_dominio': wi_dom,
+    }
+    return render(request, 'url_detalle.html', context)
 
 @login_required(login_url=reverse_lazy('login'))
 def busca(request):
